Repairs history and resend (#47)
* Repairs history and resend

- Disable attachment serialization by default
- Refactored history so EF serializable entity is contained in EF assembly
- Rebuilt resend functions

* unit test name clarifications

* Potentially fixes ci bug with resend mail

Local testing of ReSend succeeds, but on the CI server the DeliveryItem.IsResendable returns false unexpectedly
when the array of attachments is empty. Documented behavior clearly states that link All method should return true
for empty arrays, and this is the behavior when running locally. On the CI server though, this appears to be coming
back false.

Added explicit check for empty array
Modifying test invoke technique